Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
google-cloud-aiplatform==1.38.1
google-cloud-firestore==2.13.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
